お家作業の進捗

+++ ManagedDxlGame/program/game/chara/enemy_base.h
@@ -9,7 +9,7 @@ protected:
 	int e_att_;					//“G‚ÌUŒ‚—Í
 	int e_gpc_hdl_;				//“G‚Ì‰æ‘œƒnƒ“ƒhƒ‹
 	tnl::Vector3 e_pos_;		//“G‚ÌˆÊ’u
-	int e_move_speed_;			//“G‚ÌˆÚ“®‘¬“x
+	float e_move_speed_;			//“G‚ÌˆÚ“®‘¬“x
 	float e_view_angle_;		//“G‚Ì‹–ìŠp
 
 };
